requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
